technical_openai/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
pytest
httpx
ruff
requests